requirements.txt
django==4.2.5
celery==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
numpy==2.1.3
scikit-learn==1.5.2
nltk==3.9.1
python-dateutil==2.9.0
pandas==2.2.3
sentence-transformers==3.3.1